Fixed bugs when transmitting lot of bytes SPI
- The bugs : in the MFRC522_WriteByteStream function, when transmitting >= 2 bytes of data the MFRC522 module cannot sample the data fast enough
and that lead to data losses. So when the user try to send [0x93][0x20] (SEL and NVC bit frames) to request UID from the PICC, the previous data
was lost (either [0x93] or [0x20]) and it does not recognize the complete sequence so the Transceive function just timed out.
- The temporary solution : Add a small loop as a mean to delay between each SPI transmit.